+++ Sources/Sentry/SentryScope.m
@@ -169,9 +169,7 @@ SentryScope ()
     @synchronized(_fingerprintArray) {
         [_fingerprintArray removeAllObjects];
     }
-    @synchronized(_attachmentArray) {
-        [_attachmentArray removeAllObjects];
-    }
+    [self clearAttachments];
     @synchronized(_spanLock) {
         _span = nil;
     }
@@ -378,6 +376,13 @@ SentryScope ()
 {
     @synchronized(_attachmentArray) {
         [_attachmentArray addObject:attachment];
+    }
+}
+
+- (void)clearAttachments
+{
+    @synchronized(_attachmentArray) {
+        [_attachmentArray removeAllObjects];
     }
 }