+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -71,6 +71,7 @@
 #import "SentryMigrateSessionInit.h"
 #import "SentryNSError.h"
 #import "SentryNSURLRequest.h"
+#import "SentryNetworkTracker.h"
 #import "SentryOutOfMemoryLogic.h"
 #import "SentryOutOfMemoryTracker.h"
 #import "SentryPerformanceTracker.h"
@@ -107,3 +108,4 @@
 #import "SentryUIViewControllerSanitizer.h"
 #import "SentryUserFeedback.h"
 #import "TestSentryCrashAdapter.h"
+#import "URLSessionTaskMock.h"